requirements.txt
django
pandas
openpyxl
python-docx
django-bootstrap5